[chores] 리듀서 쪼개기 #39
